tonight.lua: added a lot
trying to move the colour scheme towards treesitter as much as it will
currently allow me to. this is so formatting is more universal instead
of me having to make a colour format for every language.

some things are still missing but i am working on that. right now it's
usable enough if you don't mind some theming missing.